+++ packages/flame/CHANGELOG.md
@@ -4,6 +4,7 @@
  - Fix camera not ending up in the correct position on long jumps
  - Make the `JoystickPlayer` a `PositionComponent`
  - Extract shared logic when handling components set in BaseComponent and BaseGame to ComponentSet.
+ - Fix `SpriteAnimationComponent` docs to use `Future.wait`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories