+++ packages/flame/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Extract shared logic when handling components set in BaseComponent and BaseGame to ComponentSet.
  - Rename `camera.shake(amount: x)` to `camera.shake(duration: x)`
  - Fix `SpriteAnimationComponent` docs to use `Future.wait`
+ - Add an empty `postRender` method that will run after each components render method
  - Rename `Tapable` to `Tappable`
  - Rename `HasTapableComponents` to `HasTappableComponents`