+++ packages/flame/CHANGELOG.md
@@ -8,7 +8,7 @@
  - Fix `SpriteAnimationComponent` docs to use `Future.wait`
  - Add an empty `postRender` method that will run after each components render method
  - Rename `Tapable` to `Tappable`
- - Rename `HasTapableComponents` to `HasTappableComponents`
+ - Fix `SpriteAnimationComponent` docs to use `Future.wait`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories