+++ packages/flame/CHANGELOG.md
@@ -10,6 +10,9 @@
  - Rename `Tapable` to `Tappable`
  - Fix `SpriteAnimationComponent` docs to use `Future.wait`
  - Add an empty `postRender` method that will run after each components render method
+ - Rename `HasTapableComponents` to `HasTappableComponents`
+ - Rename `prepareCanvas` to `preRender`
+ - Add `intensity` to `Camera.shake`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories