+++ packages/flame/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [Next]
+## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps
  - Make the `JoystickPlayer` a `PositionComponent`
  - Extract shared logic when handling components set in BaseComponent and BaseGame to ComponentSet.